#78 [refactor] 스타디움, 결과 테스트 코드를 수정한다.
